backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
certifi==2025.1.31
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
pydantic==2.10.2
pydantic-settings==2.6.1
web3==7.0.0
requests==2.32.0
